Partial cleanup of SpanExtensionsLab
