+++ src/System.Buffers.Primitives/System.Buffers.Primitives.csproj
@@ -12,6 +12,5 @@
     <PackageReference Include="System.Memory" Version="4.4.0-*" />
     <PackageReference Include="System.Runtime.CompilerServices.Unsafe" Version="4.4.0-*" />
     <PackageReference Include="System.Diagnostics.Tools" Version="4.3.0" />
-    <PackageReference Include="System.Numerics.Vectors" Version="4.3.0" />
   </ItemGroup>
 </Project>
