+++ src/System.IO.Pipelines.Extensions/DefaultReadableBufferExtensions.cs
@@ -1,4 +1,5 @@
-﻿using System.Numerics;
+﻿using System.Binary;
+using System.Numerics;
 using System.Runtime.CompilerServices;
 
 namespace System.IO.Pipelines
