+++ src/System.Text.Json/System/Text/Json/JsonParseObject.cs
@@ -1,6 +1,7 @@
 ﻿// Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.Binary;
 using System.Buffers;
 using System.Buffers.Pools;
 using System.Collections.Generic;
