+++ tests/System.IO.Pipelines.Tests/System.IO.Pipelines.Tests.csproj
@@ -33,6 +33,7 @@
     <ProjectReference Include="..\..\src\System.Text.Formatting\System.Text.Formatting.csproj" />
     <ProjectReference Include="..\..\src\System.Text.Primitives\System.Text.Primitives.csproj" />
     <ProjectReference Include="..\..\src\System.Buffers.Primitives\System.Buffers.Primitives.csproj" />
+    <ProjectReference Include="..\..\src\System.Binary\System.Binary.csproj" />
     <ProjectReference Include="..\..\src\System.Collections.Sequences\System.Collections.Sequences.csproj" />
   </ItemGroup>
   <!-- register for test discovery in Visual Studio -->
